scripts/revenium/pyproject.toml
[project]
name = "revenium-seed"
version = "0.1.0"
description = "Seed circuits.com demo data into a Revenium tenant (orgs, subscribers, products, backdated metering)."
requires-python = ">=3.12"
dependencies = [
  "revenium-metering>=6.8",
  "httpx>=0.28",
  "pydantic>=2.10",
  "pyyaml>=6.0",
  "python-dotenv>=1.0",
  "tenacity>=9.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[project.scripts]
revenium-seed = "revenium_seed.main:cli"

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["revenium_seed"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
